requirements.txt
absl-py==0.9.0
astor==0.8.1
Django==2.2.9
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.18.1
Pillow==7.0.0
protobuf==3.11.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.16.0
theano==1.0.4
keras==2.3.1
opencv-python==4.2.0
